Baseline for main and scanner is provided, plus unit testing
The unit testing is providing me a different problem right now, passing off everything being read (regardless if it follows the grammar or not) as 100% passing. This is NOT correct, as even the incorrect inputs should be registering as invalid rather than valid. Something is horribly wrong, and I can't figure out what in the moment.